feat: join_comfrim, board_update, board_delete

+++ SpringBoard-mybatis/src/main/webapp/WEB-INF/views/user/userlist.jsp
@@ -31,7 +31,7 @@
 							<td>${user.userPasswd}</td>
 							<td>${user.userName}</td>
 							<td>${user.userAge}</td>
-							<td><a href="/user/delete?id=${user.id}"><button>삭제</button></a></td>
+							<td><a href="/user/delete?id=${user.id}"  onclick="return confirm('정말로 삭제하시겠습니까?')"><button>삭제</button></a></td>
 						</tr>
 					</c:forEach>
 				</tbody>
